# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CK Group Spare Parts Inventory Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/monthly-expense:
    get:
      operationId: getMonthlyExpenseReport
      tags: [reports]
      summary: Monthly expense analytics
      parameters:
        - in: query
          name: year
          schema:
            type: integer
      responses:
        "200":
          description: Monthly expense report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyExpenseSummary"
  /reports/yearly-summary:
    get:
      operationId: getYearlySummary
      tags: [reports]
      summary: Yearly expense and purchase summary
      parameters:
        - in: query
          name: year
          schema:
            type: integer
      responses:
        "200":
          description: Yearly summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YearlySummary"
components:
  schemas:
...
    CategoryExpense:
      type: object
      required:
        - category
        - amount
      properties:
        category:
          type: string
        amount:
          type: number
    MonthlyExpenseSummary:
      type: object
      required:
        - month
        - year
        - totalExpense
        - categoryBreakdown
      properties:
        month:
          type: integer
        year:
          type: integer
        monthName:
          type: string
        totalExpense:
          type: number
        previousMonthExpense:
          type: number
          nullable: true
        growthPercentage:
          type: number
          nullable: true
        highestCategory:
          type: string
          nullable: true
        lowestCategory:
          type: string
          nullable: true
        categoryBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/CategoryExpense"
    YearlySummary:
      type: object
      required:
        - year
        - totalAnnualExpense
        - highestSpendingMonth
        - lowestSpendingMonth
        - mostExpensiveCategory
        - totalPurchasedParts
      properties:
        year:
          type: integer
        totalAnnualExpense:
          type: number
        highestSpendingMonth:
          type: string
        lowestSpendingMonth:
          type: string
        mostExpensiveCategory:
          type: string
        totalPurchasedParts:
          type: integer