# lib/api-zod/src/generated/types/categoryExpense.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CK Group Spare Parts Inventory Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryExpense {
  category: string;
  amount: number;
}